Fix tests for targets with sizeof(int) != 32.
gcc/testsuite/

	* gcc.dg/torture/pr69352.c (foo): Cast to intptr_t instead of long.
	* gcc.dg/torture/pr69771.c: Require int32plus.
	* gcc.dg/torture/pr71866.c (inb): Add cast to intptr_t.

From-SVN: r238708

+++ gcc/testsuite/gcc.dg/torture/pr69352.c
@@ -1,5 +1,7 @@
 /* { dg-do compile } */
 
+#include <stdint.h>
+
 int a[10][14], b, c, d, e, f, g, h, i;
 void bar (void);
 int
@@ -13,7 +15,7 @@ foo (int x)
   else
     m = 13;
   if (a[x][m])
-    l = (long) foo;
+    l = (intptr_t) foo;
   a[x][i] = l;
   while (c)
     {
